app: split up preference items
