+++ app/app/src/main/java/de/stustanet/stustapay/ui/pref/PrefLink.kt
@@ -1,0 +1,47 @@
+package de.stustanet.stustapay.ui.pref
+
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.*
+import androidx.compose.material.Divider
+import androidx.compose.material.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+
+@Composable
+fun PrefLink(
+    modifier: Modifier = Modifier,
+    icon: (@Composable () -> Unit)? = null,
+    title: @Composable () -> Unit,
+    subtitle: (@Composable () -> Unit)? = null,
+    action: (@Composable () -> Unit)? = null,
+    onClick: () -> Unit,
+) {
+    Surface {
+        Row(
+            modifier = modifier.fillMaxWidth(),
+        ) {
+            Row(
+                modifier = Modifier
+                    .weight(1f)
+                    .clickable(onClick = onClick),
+                verticalAlignment = Alignment.CenterVertically,
+            ) {
+                CenterIcon(icon = icon)
+                PrefTitle(title = title, subtitle = subtitle)
+            }
+            if (action != null) {
+                Divider(
+                    modifier = Modifier
+                        .padding(vertical = 4.dp)
+                        .height(56.dp)
+                        .width(1.dp),
+                )
+                CenterPadContent {
+                    action()
+                }
+            }
+        }
+    }
+}
